Renaming interfaces and removing I. Introducing DynamoDBCheckpointer.
