+++ amazon-kinesis-client/src/main/java/software/amazon/kinesis/leases/LeaseManagementFactory.java
@@ -23,7 +23,7 @@ public interface LeaseManagementFactory {
 
     ShardSyncTaskManager createShardSyncTaskManager();
 
-    LeaseManager<KinesisClientLease> createLeaseManager();
+    DynamoDBLeaseManager<KinesisClientLease> createLeaseManager();
 
     KinesisClientLibLeaseCoordinator createKinesisClientLibLeaseCoordinator();
