+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/coordinator/SchedulerTest.java
@@ -50,13 +50,13 @@ import com.amazonaws.services.kinesis.clientlibrary.exceptions.KinesisClientLibN
 import software.amazon.kinesis.checkpoint.Checkpoint;
 import software.amazon.kinesis.checkpoint.CheckpointConfig;
 import software.amazon.kinesis.checkpoint.CheckpointFactory;
-import software.amazon.kinesis.leases.ILeaseManager;
+import software.amazon.kinesis.leases.LeaseManager;
 import software.amazon.kinesis.leases.KinesisClientLease;
 import software.amazon.kinesis.leases.KinesisClientLibLeaseCoordinator;
 import software.amazon.kinesis.leases.LeaseCoordinator;
 import software.amazon.kinesis.leases.LeaseManagementConfig;
 import software.amazon.kinesis.leases.LeaseManagementFactory;
-import software.amazon.kinesis.leases.LeaseManager;
+import software.amazon.kinesis.leases.DynamoDBLeaseManager;
 import software.amazon.kinesis.leases.LeaseManagerProxy;
 import software.amazon.kinesis.leases.ShardInfo;
 import software.amazon.kinesis.leases.ShardSyncTaskManager;
@@ -111,7 +111,7 @@ public class SchedulerTest {
     @Mock
     private ShardSyncTaskManager shardSyncTaskManager;
     @Mock
-    private LeaseManager<KinesisClientLease> leaseManager;
+    private DynamoDBLeaseManager<KinesisClientLease> dynamoDBLeaseManager;
     @Mock
     private LeaseManagerProxy leaseManagerProxy;
     @Mock
@@ -131,7 +131,7 @@ public class SchedulerTest {
         processorConfig = new ProcessorConfig(processorFactory);
         retrievalConfig = new RetrievalConfig(streamName, amazonKinesis).retrievalFactory(retrievalFactory);
 
-        when(leaseCoordinator.leaseManager()).thenReturn(leaseManager);
+        when(leaseCoordinator.leaseManager()).thenReturn(dynamoDBLeaseManager);
         when(shardSyncTaskManager.leaseManagerProxy()).thenReturn(leaseManagerProxy);
         when(retrievalFactory.createGetRecordsCache(any(ShardInfo.class), any(IMetricsFactory.class))).thenReturn(getRecordsCache);
 
@@ -442,8 +442,8 @@ public class SchedulerTest {
         }
 
         @Override
-        public LeaseManager<KinesisClientLease> createLeaseManager() {
-            return leaseManager;
+        public DynamoDBLeaseManager<KinesisClientLease> createLeaseManager() {
+            return dynamoDBLeaseManager;
         }
 
         @Override
@@ -460,7 +460,7 @@ public class SchedulerTest {
     private class TestKinesisCheckpointFactory implements CheckpointFactory {
         @Override
         public Checkpointer createCheckpointer(final LeaseCoordinator<KinesisClientLease> leaseCoordinator,
-                                               final ILeaseManager<KinesisClientLease> leaseManager) {
+                                               final LeaseManager<KinesisClientLease> leaseManager) {
             return checkpoint;
         }
     }
